Use parameterized to enable parallelisation of slow tests
We can't use pytest's parameterisation here as this test is a
unittest-style test, to which pytest does not support applying
parameterisation. Instead we can use the separate parameterized
package which offers parameterisation on plain unittest tests,
including when run under pytest.

Happily this also plays nicely with pytest-xdist. As the resulting
test functions are actually separate functions on the class, xdist
is able to then spread the work between workers.